+++ src/StkTable/index.ts
@@ -1,4 +1,4 @@
 export { default as StkTable } from './StkTable.vue';
-export { tableSort, insertToOrderedArray, strCompare } from './utils';
+export { tableSort, insertToOrderedArray, strCompare, binarySearch } from './utils';
 export type { StkTableColumn } from './types/index';
 import './style.less';