+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.6.6
+* Feature
+  - export `binarySearch` function.
+  - add: `customBottom` slots.
+  
 ## 0.6.5
 * Feature
   - support expand row in virtual mode.